fixxx

+++ backend/DTOs/ProductDto.cs
@@ -4,5 +4,5 @@
 
 namespace CarTechShowcase.DTOs;
 
-public record ProductCreateDto(string ProductName, decimal? Price, string? Category, string? Description, string? Img, string? Spec);
-public record ProductUpdateDto(string? ProductName, decimal? Price, string? Category, string? Description, string? Img, string? Spec, bool? IsActive);
+public record ProductCreateDto(string ProductName, double? Price, string? Category, string? Description, string? Img, string? Spec);
+public record ProductUpdateDto(string? ProductName, double? Price, string? Category, string? Description, string? Img, string? Spec, bool? IsActive);
